Viz - Fix cyclical dependencies in axes
