+++ js/viz/axes/polar_axes.js
@@ -7,7 +7,7 @@ import { extend } from '../../core/utils/extend';
 import constants from './axes_constants';
 import xyAxes from './xy_axes';
 import { tick } from './tick';
-import { calculateCanvasMargins, Axis, measureLabels } from './base_axis';
+import { calculateCanvasMargins, measureLabels } from './axes_utils';
 import { noop as _noop } from '../../core/utils/common';
 const { PI, abs, atan, round } = Math;
 const _min = Math.min;
@@ -156,15 +156,13 @@ const circularAxes = {
         return margins;
     },
 
-    updateSize() {
+    _updateLabelsPosition() {
         const that = this;
-
-        Axis.prototype.updateSize.apply(that, arguments);
 
         measureLabels(that._majorTicks);
         that._adjustLabelsCoord(0, 0, true);
 
-        this._checkBoundedLabelsOverlapping(this._majorTicks, this._majorTicks.map(t=>t.labelBBox));
+        that._checkBoundedLabelsOverlapping(this._majorTicks, this._majorTicks.map(t=>t.labelBBox));
     },
 
     _setVisualRange: _noop,
